move ListVoteCount helper methods to AgendaHelper
